Upload library updates

+++ IOSMatchingEngineSDK/Example/Tests/Tests.swift
@@ -1,4 +1,4 @@
-// Copyright 2020 MobiledgeX, Inc. All rights and licenses reserved.
+// Copyright 2018-2020 MobiledgeX, Inc. All rights and licenses reserved.
 // MobiledgeX, Inc. 156 2nd Street #408, San Francisco, CA 94105
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -25,7 +25,7 @@ class Tests: XCTestCase {
     let TEST = true
     
     var host = ""
-    var port: UInt = 38001
+    var port: UInt16 = 38001
     var appName: String!
     var appVers: String!
     var devName: String!
