+++ IOSMatchingEngineSDK/MobiledgeXiOSLibrary/Classes/MatchingEngine/AppInstList.swift
@@ -1,4 +1,4 @@
-// Copyright 2020 MobiledgeX, Inc. All rights and licenses reserved.
+// Copyright 2018-2020 MobiledgeX, Inc. All rights and licenses reserved.
 // MobiledgeX, Inc. 156 2nd Street #408, San Francisco, CA 94105
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -117,7 +117,7 @@ extension MobiledgeXiOSLibrary.MatchingEngine {
         return getAppInstList(host: host, port: port, request: request)
     }
     
-    public func getAppInstList(host: String, port: UInt, request: [String: Any])
+    public func getAppInstList(host: String, port: UInt16, request: [String: Any])
         -> Promise<[String: AnyObject]>
     {
         os_log("Finding nearby appInsts matching this MatchingEngine client.", log: OSLog.default, type: .debug)
