+++ IOSMatchingEngineSDK/MobiledgeXiOSLibrary/Classes/MatchingEngine/GetLocation.swift
@@ -1,5 +1,5 @@
 
-// Copyright 2020 MobiledgeX, Inc. All rights and licenses reserved.
+// Copyright 2018-2020 MobiledgeX, Inc. All rights and licenses reserved.
 // MobiledgeX, Inc. 156 2nd Street #408, San Francisco, CA 94105
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -109,7 +109,7 @@ extension MobiledgeXiOSLibrary.MatchingEngine {
     ///   - port: port override of the dme server port
     ///   - request: GetLocationRequest dictionary, from createGetLocationRequest.
     /// - Returns: API Dictionary/json
-    public func getLocation(host: String, port: UInt, request: [String: Any]) -> Promise<[String: AnyObject]>
+    public func getLocation(host: String, port: UInt16, request: [String: Any]) -> Promise<[String: AnyObject]>
     {
         let promiseInputs: Promise<[String: AnyObject]> = Promise<[String: AnyObject]>.pending()
         os_log("getLocation", log: OSLog.default, type: .debug)
