+++ IOSMatchingEngineSDK/MobiledgeXiOSLibrary/Classes/MatchingEngine/RegisterClient.swift
@@ -1,4 +1,4 @@
-// Copyright 2020 MobiledgeX, Inc. All rights and licenses reserved.
+// Copyright 2018-2020 MobiledgeX, Inc. All rights and licenses reserved.
 // MobiledgeX, Inc. 156 2nd Street #408, San Francisco, CA 94105
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -146,7 +146,7 @@ extension MobiledgeXiOSLibrary.MatchingEngine
     ///   - port: port override of the dme server port
     ///   - request: RegisterClient dictionary, from createRegisterClientReqwuest.
     /// - Returns: API Dictionary/json
-    public func registerClient(host: String, port: UInt, request: [String: Any]) -> Promise<[String: AnyObject]>
+    public func registerClient(host: String, port: UInt16, request: [String: Any]) -> Promise<[String: AnyObject]>
     {
         let promiseInputs: Promise<[String: AnyObject]> = Promise<[String: AnyObject]>.pending()
         os_log("registerClient", log: OSLog.default, type: .debug)
